pyproject.toml
[project]
name = "user_service"
version = "0.1.0"
description = "Modüler FastAPI mikroservisi"
authors = [
    { name = "Mübarek", email = "0xmillennium@protonmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"

# `pip install .` komutu çalıştığında kurulacak ana bağımlılıklar
dependencies = [
    "pydantic[email] ~=2.8.2",
    "fastapi ~=0.112.2",
    "sqlalchemy ~=2.0.34",
    "aio-pika ~=9.5.5",
    "starlette ~=0.38.2",
    "aiormq ~=6.8.1",
    "PyYAML ~=6.0.1",
    "uvicorn ~=0.34.2",
    "argon2-cffi ~=21.3.0",
    "psycopg2 ~=2.9.10"
]

[project.optional-dependencies]
dev = [
    "pytest ~=8.3.5"
]

[project.urls]
Homepage = "https://github.com/your/repo"
"Repository" = "soon..."

[tool.setuptools]
# setuptool ile kurulum yaparken kaynak kodu buradan al
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
